feat(auth): move signin and signup to be root level outside of /users
Moved signin and signup pages to be root level so that the route is just /signin or /signup. Also
moved Auth.module.css to root level of styles to match route. Added restricted page control with
session on signin and signup pages.
